Refactor: BottomSheet common → ui, Codeit 슬롯 API 도입
- 위치 일관: components/common/BottomSheet/ → components/ui/BottomSheet/
- Codeit 4-section 슬롯: title? / ariaLabel? / showClose / footer / children
- a11y: trim 기반 accessibleLabel + dev mode 빈 문자열 경고 (footgun 방지)
- footer slot — AdvancedFilterSheet 자체 footer 제거 → BottomSheet 위임 (적용/초기화 Button)
- 4 caller import 경로 업데이트 (CategoryBottomSheet, SeriesBrowserSheet, AdvancedFilterSheet, admin/FilterDropdown)
- close_btn margin-left:auto 로 title 없을 때도 우측 정렬
