+++ src/app/(content)/sermons/_component/AdvancedFilterSheet/AdvancedFilterSheet.module.scss
@@ -45,10 +45,3 @@
   }
 }
 
-.footer {
-  margin-top: $content-gap-l;
-  display: flex;
-  gap: $spacing-12;
-  padding-top: $spacing-16;
-  border-top: 1px solid $border-primary;
-}
